backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
pydantic>=2.0.0
tensorflow>=2.10.0
numpy>=1.23.0
Pillow>=9.0.0
python-dotenv>=1.0.0
